포인트 조회api

+++ src/main/java/com/cdy/ecommerce/ecommerce/admin/ProductAdminDTO.java
@@ -18,7 +18,7 @@ public class ProductAdminDTO {
         private String userId; // 상품을 등록하는 사용자의 ID
         private String name; // 상품 이름
         private int price; // 상품 가격
-        private Long initialStock; // 초기 재고 수량
+        private int initialStock; // 초기 재고 수량
     }
 
     @Data
@@ -29,6 +29,6 @@ public class ProductAdminDTO {
         private Long id; // 상품 ID
         private String name; // 상품 이름
         private int price; // 상품 가격
-        private Long stock; // 현재 재고 수량
+        private int stock; // 현재 재고 수량
     }
 }
