+++ src/main/java/com/cdy/ecommerce/ecommerce/domain/product/business/models/ProductStock.java
@@ -17,16 +17,16 @@ public class ProductStock {
     @JoinColumn(name = "product_id")
     private Product product;
 
-    private Long quantity;
+    private int quantity;
 
     public ProductStock(){}
 
-    public ProductStock(Product product, Long quantity){
+    public ProductStock(Product product, int quantity){
         this.product = product;
         this.quantity = quantity;
     }
 
-    public void decrease(Long quantity){
+    public void decrease(int quantity){
         if(this.quantity - quantity < 0){
             throw new StockException("재고는 0개 미만이 될 수 없습니다.");
         }
